requirement.txt
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
cffi==1.10.0
colorama==0.3.7
configparser==3.5.0
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
diff-match-patch==20121119
Django==1.10.5
django-bootstrap-form==3.2.1
django-el-pagination==3.0.1
django-import-export==0.5.1
django-pagination==1.0.7
django-registration-redux==1.4
django-report-builder==3.4.4
django-tables2==1.4.2
djangorestframework==3.5.4
et-xmlfile==1.0.1
httpie==0.9.9
idna==2.5
incremental==16.10.1
jdcal==1.3
olefile==0.44
openpyxl==2.3.5
packaging==16.8
Pillow==4.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
Pygments==2.2.0
PyGoogleVoice==0.5.2
PyMySQL==0.7.9
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
service-identity==16.0.0
six==1.10.0
tablib==0.11.3
zope.interface==4.3.3
